server/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pillow>=10.0.0
numpy>=1.24.0
loguru>=0.7.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
